fix using same formKey issue
This fixes the issue when viewing a notice or creating a notice using the same form key, causing an error in the widget tree because all these forms are watched from the same RiverPod provider.
By using provider scope, then overriding with a copy, fix this issue.
In this commit, also move the view notice route outside the immediate shell route because a direct subroute of a shell route cannot have a different parent key.
Doing this help view the notice route live on the root navigation stack and cover the navigation rail/bar.